Add test_coverage
